backend/requirements.txt
# --- Web (FastAPI) ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary>=2.9.10
alembic==1.13.3
pgvector==0.3.5

# --- Celery / cache ---
celery[redis]==5.4.0
redis==5.1.1

# --- AI ---
openai==1.51.0

# --- HTTP / upload ---
httpx==0.27.2
python-multipart==0.0.12

# --- OCR / images ---
google-cloud-vision==3.7.4
pytesseract==0.3.13
Pillow==10.4.0
imagehash==4.3.1

# --- Storage ---
cloudinary==1.41.0

# --- Google Sheets ---
gspread==6.1.2
google-auth==2.35.0

# --- Auth ---
PyJWT==2.9.0
passlib[bcrypt]==1.7.4

# --- Clustering / analytics (workers) — cài riêng khi cần Celery clustering: pip install -r requirements-workers.txt
# (trên Windows/Python 3.14 thường cần wheel sẵn hoặc Visual Studio Build Tools)

# --- Utils ---
python-dateutil==2.9.0
tenacity==9.0.0
certifi==2024.8.30
